move sync impl into sync module
