+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "elastic_reqwest"
-version = "0.8.0"
+version = "0.9.0"
 authors = ["Ashley Mannix <ashleymannix@live.com.au>", "Stephan Buys <stephan.buys@gmail.com>"]
 license = "Apache-2.0"
 description = "A lightweight implementation of the Elasticsearch API based on reqwest."
@@ -14,8 +14,8 @@ nightly = []
 [dependencies]
 serde = "~1"
 elastic_requests = "~0.4.1"
-elastic_responses = "~0.8.0"
-reqwest = "~0.6.0"
+elastic_responses = "~0.9.0"
+reqwest = { version = "*", git = "https://github.com/seanmonstar/reqwest.git" }
 url = "~1"
 
 [dev-dependencies]
