+++ examples/typed/Cargo.toml
@@ -10,7 +10,7 @@ reqwest = "~0.6"
 serde = "~1"
 serde_derive = "~1"
 serde_json = "~1"
-elastic_types = "~0.18"
-elastic_types_derive = "~0.18"
+elastic_types = "~0.19"
+elastic_types_derive = "~0.19"
 
 json_str = { version = "*", features = ["nightly"]}
